+++ Sources/MapboxMaps/ContentBuilders/Tree/MapContentReconciler.swift
@@ -77,6 +77,7 @@ private extension MapContentNodeContext {
         uniqueProperties.update(
             from: oldProperties,
             style: style.styleManager,
+            initial: initialUniqueProperties,
             locationManager: content?.location.value
         )
     }
@@ -85,19 +86,37 @@ private extension MapContentNodeContext {
         lastLayerId = nil
 
         /// On style reload we need to traverse the whole tree to reconstruct non-persistent layers
-        /// On style we need to identifty bottom position in the style in order to stack content above
+        /// On style reload we need to identify the bottom position in the style in order to stack content above
         /// Position must take into account only non-persistent layers, which was not added in runtime
         isEqualContent = { _, _ in false }
         initialStyleLayers = getInitialStyleLayers() ?? []
         initialStyleImports = style.styleManager.getStyleImports().map(\.id)
+        initialUniqueProperties = getInitialUniqueProperties()
+
         mapContent.update(root)
         isEqualContent = arePropertiesEqual
 
         uniqueProperties.update(
             from: MapContentUniqueProperties(),
             style: style.styleManager,
+            initial: initialUniqueProperties,
             locationManager: content?.location.value
         )
+    }
+
+    func getInitialUniqueProperties() -> MapContentUniqueProperties? {
+        if let jsonData = style.styleManager.getStyleJSON().data(using: .utf8) {
+            do {
+                var initialMapUniqueProperties = try JSONDecoder().decode(MapContentUniqueProperties.self, from: jsonData)
+                // Transition options are not included in the StyleJSON
+                initialMapUniqueProperties.transition = TransitionOptions(style.styleManager.getStyleTransition())
+                return initialMapUniqueProperties
+            } catch {
+                Log.warning(forMessage: "Unable to decode initial MapContentUniqueProperties \(error) from StyleJSON", category: "StyleDSL")
+                return nil
+            }
+        }
+        return nil
     }
 
     func getInitialStyleLayers() -> [String]? {